backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Zoom-Developer <sa-mp12345@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
sqlmodel = "^0.0.22"
uvicorn = "^0.34.0"
gunicorn = "^23.0.0"
alembic = "^1.14.1"
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"
aioboto3 = "^13.4.0"
python-multipart = "^0.0.20"
pillow = "^11.1.0"
openai = "^1.62.0"
prometheus-fastapi-instrumentator = "^7.0.2"
redis = "^5.2.1"


[tool.poetry.group.dev.dependencies]
testcontainers = "^4.9.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
minio = "^7.2.15"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "./"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    'ignore::DeprecationWarning'
]

